chore(polling) Define non-specialized polling model
